+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Bugs fixed
 
 * Fix `set_pcxxheader` bug
+* [#140](https://github.com/tboox/xmake/issues/140): Fix `os.tmpdir()` in fakeroot
 
 ## v2.1.6
 
@@ -357,6 +358,7 @@
 ### Bugs修复
 
 * 修复`set_pcxxheader`编译没有继承flags配置问题
+* [#140](https://github.com/tboox/xmake/issues/140): 修复`os.tmpdir()`在fakeroot下的冲突问题
 
 ## v2.1.6